+++ packages/gov-ui-kit/src/modules/components/member/index.ts
@@ -1,1 +1,2 @@
 export * from './memberAvatar';
+export * from './memberDataListItem';