+++ packages/gov-ui-kit/src/modules/components/member/memberDataListItem/memberDataListItemStructure/index.ts
@@ -1,0 +1,7 @@
+import { MemberDataListItemStructure } from './memberDataListItemStructure';
+
+export const MemberDataListItem = {
+    Structure: MemberDataListItemStructure,
+};
+
+export type { IMemberDataListItemProps } from './memberDataListItemStructure';